Move test cases to separate class
